pyproject.toml
[project]
name = "owl-cli"
version = "0.1.3"
description = "Semantic code search using vector embeddings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "sentence-transformers>=3.0",
    "torch>=2.0",
    "faiss-cpu>=1.7",
    "numpy>=1.26",
    "tree-sitter>=0.22",
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
    "tree-sitter-java>=0.23",
    "tree-sitter-go>=0.23",
    "tree-sitter-ruby>=0.23",
    "tree-sitter-rust>=0.23",
    "tree-sitter-php>=0.23",
    "pathspec>=0.12",
    "mcp>=1.0",
    "tqdm>=4.66",
    "rich>=13.0",
]

[project.scripts]
owl = "owl_cli.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
